requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymupdf
faiss-cpu
numpy
google-generativeai
pydantic
pdfplumber